fix: buttonVariants in ServerPage
