+++ ui/button/button-variants.ts
@@ -1,11 +1,4 @@
-'use client'
-
-import * as React from 'react'
-import { Slot } from '@radix-ui/react-slot'
-import { cva, type VariantProps } from 'class-variance-authority'
-
-import { cn } from '@/lib/utils'
-
+import { cva } from 'class-variance-authority'
 // Shared style patterns for consistency and reusability
 const interactiveStates = {
   primary: 'hover:bg-primary-hover focus:bg-primary-focus active:bg-primary-pressed',
@@ -45,7 +38,7 @@ const iconButtonStyles = `bg-transparent text-iconcolor-primary hover:bg-backgro
 
 const volumeButtonStyles = `px-6 py-4 bg-primary text-primary-foreground ${interactiveStates.primary} ${focusRing.primary} shadow-md [background-image:linear-gradient(45deg,rgba(31,210,192,0)_60%,rgba(31,210,192,0.5)_80%,rgba(255,255,255,0.5)_100%)] [box-shadow:inset_0px_2px_4px_rgba(255,255,255,0.3)] [-webkit-font-smoothing:antialiased] [-moz-osx-font-smoothing:grayscale] disabled:[background-image:none]`
 
-const buttonVariants = cva(
+export const buttonVariants = cva(
   'inline-flex items-center justify-center whitespace-normal rounded-full text-base leading-none transition-colors focus:outline-none disabled:pointer-events-none disabled:bg-background-muted disabled:text-textcolor-muted disabled:[text-shadow:0_1px_0_hsla(0,0%,100%,0.8)] [&_svg]:pointer-events-none [&_svg]:shrink-0 h-10',
   {
     variants: {
@@ -96,18 +89,3 @@ const buttonVariants = cva(
     ],
   }
 )
-
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof buttonVariants> {
-  asChild?: boolean
-}
-
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-  ({ className, variant, size, asChild = false, ...props }, ref) => {
-    const Comp = asChild ? Slot : 'button'
-    return <Comp className={cn(buttonVariants({ variant, size, className }))} ref={ref} {...props} />
-  }
-)
-Button.displayName = 'Button'
-
-export { Button, buttonVariants }
